Abstract SicpControlBar into its own component
